Fixed HTTP Request and controller handler for deletecomment request, settime out for a refresh on viewpost component to re-pull the viewpost data so the comment disappears after hitting the delete button for the comment

+++ beatsPM/src/main/java/com/example/beatsPM/Models/CommentModel.java
@@ -17,7 +17,7 @@ public class CommentModel {
 //    @ManyToOne(fetch = LAZY)
 //    @JoinColumn(name = "postId", referencedColumnName = "postId")
     @NotNull
-    private Integer postId;
+    private Long postId;
 
 
     @GeneratedValue
@@ -39,11 +39,11 @@ public class CommentModel {
         this.commentBody = commentBody;
     }
 
-    public int getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(int postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
